refactor: renamed cw721_utils to cw721_test_utils and made compile only during test
jira: DCD-50
